refactor: add request, response Parameter DTO to UserService
