--- src/main/java/com/sprint/mission/discodeit/entity/BinaryContentCreateRequest.java
@@ -1,11 +1,0 @@
-package com.sprint.mission.discodeit.entity;
-
-import java.util.UUID;
-
-public record BinaryContentCreateRequest(
-        UUID userId,
-        UUID messageId,
-        String fileName,
-        String fileType
-) {
-}
